finished invite to server modal

+++ client/src/components/servers/DeleteServer.js
@@ -1,13 +1,26 @@
 import { useContext, useState } from 'react';
 import Modal from 'react-modal';
 import ServerContext from '../../context/serverContext';
-import { DELETE_SERVER, OPEN_MODAL } from '../../const';
+import {
+  DELETE_SERVER,
+  OPEN_MODAL,
+  DEFAULT_SERVER,
+  GET_SERVER,
+} from '../../const';
 import { useMutation, gql } from '@apollo/client';
 import { customStyles } from '../channels/CreateChannelForm';
 
 const DELETE_SERVER_MUTATION = gql`
   mutation DeleteServer($serverId: ID!) {
     deleteServer(serverId: $serverId) {
+      id
+    }
+  }
+`;
+
+const LEAVE_SERVER_MUTATION = gql`
+  mutation LeaveServer($serverId: ID!) {
+    leaveServer(serverId: $serverId) {
       id
     }
   }
@@ -29,27 +42,37 @@ const DeleteServer = () => {
 
   const [deleteServer] = useMutation(DELETE_SERVER_MUTATION, {
     update(cache, { data: { deleteServer } }) {
-      cache.modify({
-        id: cache.identify(deleteServer),
-        fields: {
-          userServers(_, { DELETE }) {
-            return DELETE;
-          },
-        },
-      });
+      const serverToDelete = cache.identify(deleteServer);
+      cache.evict({ id: serverToDelete });
+    },
+  });
+
+  const [leaveServer] = useMutation(LEAVE_SERVER_MUTATION, {
+    update(cache, { data: { leaveServer } }) {
+      const serverToDelete = cache.identify(leaveServer);
+      cache.evict({ id: serverToDelete });
     },
   });
 
   const handleSubmit = (e) => {
     e.preventDefault();
     if (serverName === nameRef.value) {
-      deleteServer({
-        variables: {
-          serverId,
-        },
-      });
+      if (serverRole === 'ADMIN') {
+        deleteServer({
+          variables: {
+            serverId,
+          },
+        });
+      } else {
+        leaveServer({
+          variables: {
+            serverId,
+          },
+        });
+      }
       setAlert(false);
       nameRef.value = '';
+      dispatch({ type: GET_SERVER, payload: DEFAULT_SERVER });
       dispatch({ type: OPEN_MODAL, payload: null });
     }
     setAlert(true);
@@ -98,10 +121,7 @@ const DeleteServer = () => {
           </div>
         </div>
         <div className='buttons-wrapper'>
-          <button
-            className='button-cancel'
-            onClick={handleClose}
-          >
+          <button className='button-cancel' onClick={handleClose}>
             Cancel
           </button>
           <button className='button-delete-server' form='delete-server'>
